Change type of nInterleaved to unsigned int

+++ include/recoil/rans_encoder.h
@@ -14,7 +14,7 @@
 namespace Recoil {
     template<UnsignedType RansStateType, UnsignedType RansBitstreamType,
             BitCountType ProbBits, RansStateType RenormLowerBound, BitCountType WriteBits,
-            size_t NInterleaved,
+            unsigned int NInterleaved,
             bool RecordIntermediateStates = false, UnsignedType RansIntermediateStateType = RansStateType>
     class RansEncoder {
         using MyRans = Rans<RansStateType, RansBitstreamType, ProbBits, RenormLowerBound, WriteBits>;
